refactor(redux/settings): convert to slice
