+++ assets/scripts/store/actions/index.js
@@ -47,9 +47,6 @@ export const SET_INFO_BUBBLE_MOUSE_INSIDE = 'SET_INFO_BUBBLE_MOUSE_INSIDE'
 export const SHOW_DESCRIPTION = 'SHOW_DESCRIPTION'
 export const HIDE_DESCRIPTION = 'HIDE_DESCRIPTION'
 
-/* settings */
-export const SET_USER_SETTINGS = 'SET_USER_SETTINGS'
-
 /* street */
 export const REPLACE_STREET_DATA = 'REPLACE_STREET_DATA'
