+++ assets/scripts/store/actions/settings.js
@@ -1,15 +1,6 @@
-import { SET_USER_SETTINGS } from './index'
+import { updateSettings } from '../slices/settings'
 import { saveSettingsLocally } from '../../users/settings'
 
-// Normal action creator.
-export function updateSettings (settings) {
-  return {
-    type: SET_USER_SETTINGS,
-    settings
-  }
-}
-
-// Thunk version of above.
 export function setSettings (settings) {
   return (dispatch) => {
     dispatch(updateSettings(settings))
