+++ assets/scripts/store/reducers/index.js
@@ -9,7 +9,7 @@ import infoBubble from './infoBubble'
 import locale from '../slices/locale'
 import map from '../slices/map'
 import menus from '../slices/menus'
-import settings from './settings'
+import settings from '../slices/settings'
 import persistSettings from '../slices/persistSettings'
 import street from './street'
 import system from '../slices/system'
