--- assets/scripts/store/selectors/settings.js
@@ -1,2 +1,0 @@
-
-export const getLastStreetId = (state) => state.settings.priorLastStreetId
