+++ assets/scripts/streets/constants.js
@@ -1,3 +1,6 @@
+export const NEW_STREET_DEFAULT = 1
+export const NEW_STREET_EMPTY = 2
+
 export const MIN_CUSTOM_STREET_WIDTH = 10
 export const MAX_CUSTOM_STREET_WIDTH = 400
